+++ tests/unit/plugins/modules/utils.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright: Josh Williams <jdubz@dubzland.net>
+# Copyright: Josh Williams <jdubz@dubzland.com>
 # GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -9,13 +9,12 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 import json
-import unittest
 
-from contextlib import contextmanager
-
-from unittest.mock import patch
+from ansible_collections.dubzland.minio.tests.unit.compat import unittest
+from ansible_collections.dubzland.minio.tests.unit.compat.mock import patch
 from ansible.module_utils import basic
 from ansible.module_utils.common.text.converters import to_bytes
+from contextlib import contextmanager
 
 
 class AnsibleExitJson(Exception):