+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ansible module `minio_alias` for managing `mc` aliases (#1)
+- Ansible module `minio_policy` for managing Policies (#4)
 
 ### Changed
 
-- Refactored testing to fall more in line with a standard collection.
+- Refactored testing to fall more in line with a standard collection. (#6)
 
 ## [0.0.1] - 2023-11-15
 
@@ -21,4 +22,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minio Client and Server role
 
-[0.0.1]: https://git.dubzland.net/dubzland/ansible-collection-minio/-/tree/0.0.1
+[0.0.1]: https://git.dubzland.com/dubzland/ansible-collection-minio/-/tree/0.0.1